Add Report_FlagOverTime.
This is a relatively simple baseline, without many deep changes from the initial version you implemented. I split things up to make it easier (I think) to follow what's happening in the code.

At the moment the test is simple and covers the overall output. If we reuse any of the unexported functions, it might make more sense to test them individually (and, long-term, I might do that anyway to make it easier to track down bugs).

This version is intentionally not interactive, to establish our baseline chart. I also have not yet integrated it into the overall report.

We should probably wait for #1629 and #1632 before finalizing this.