Fix 1606 (#1638)
* Don't fetch-depth.

fetch-depth: 0 gets the full git history for every branch. We shouldn't have to do that.

* Build dev pkgdown.

Closes #1606.

I *think* this will work, but we can't fully test until this is merged into dev, at which point we *might* break the existing site. In theory we should be able to trigger the action on main to un-break it, though.

* Separate PR pkgdown check from dev pkgdown builder.

* Rename pkgdown workflows to better match their functions.